+++ designer/src/api/toolbox-api.ts
@@ -1,11 +1,10 @@
 import { Step } from '../definition';
 import { BehaviorController } from '../behaviors/behavior-controller';
-import { ObjectCloner, Uid, Vector } from '../core';
-import { StepDefinition } from '../designer-configuration';
+import { ObjectCloner, SimpleEventListener, Uid, Vector } from '../core';
+import { StepDefinition, ToolboxConfiguration, ToolboxGroupConfiguration, UidGenerator } from '../designer-configuration';
 import { DesignerState } from '../designer-state';
 import { DragStepBehavior } from '../behaviors/drag-step-behavior';
 import { DesignerContext } from '../designer-context';
-import { LayoutController } from '../layout-controller';
 import { IconProvider } from '../core/icon-provider';
 
 export class ToolboxApi {
@@ -13,16 +12,43 @@ export class ToolboxApi {
 		private readonly state: DesignerState,
 		private readonly designerContext: DesignerContext,
 		private readonly behaviorController: BehaviorController,
-		private readonly layoutController: LayoutController,
-		private readonly iconProvider: IconProvider
+		private readonly iconProvider: IconProvider,
+		private readonly configuration: ToolboxConfiguration | false,
+		private readonly uidGenerator: UidGenerator | undefined
 	) {}
 
-	public isVisibleAtStart(): boolean {
-		return this.layoutController.isMobile();
+	public isCollapsed(): boolean {
+		return this.state.isToolboxCollapsed;
+	}
+
+	public toggleIsCollapsed() {
+		this.state.setIsToolboxCollapsed(!this.state.isToolboxCollapsed);
+	}
+
+	public subscribeIsCollapsed(listener: SimpleEventListener<boolean>) {
+		this.state.onIsToolboxCollapsedChanged.subscribe(listener);
 	}
 
 	public tryGetIconUrl(step: StepDefinition): string | null {
 		return this.iconProvider.getIconUrl(step);
+	}
+
+	public getLabel(step: StepDefinition): string {
+		const labelProvider = this.getConfiguration().labelProvider;
+		return labelProvider ? labelProvider(step) : step.name;
+	}
+
+	public filterGroups(filter: string | undefined): ToolboxGroupConfiguration[] {
+		return this.getConfiguration()
+			.groups.map(group => {
+				return {
+					name: group.name,
+					steps: group.steps.filter(s => {
+						return filter ? s.name.toLowerCase().includes(filter) : true;
+					})
+				};
+			})
+			.filter(group => group.steps.length > 0);
 	}
 
 	/**
@@ -32,16 +58,23 @@ export class ToolboxApi {
 	 */
 	public tryDrag(position: Vector, step: StepDefinition): boolean {
 		if (!this.state.isReadonly) {
-			const newStep = createStep(step);
+			const newStep = this.activateStep(step);
 			this.behaviorController.start(position, DragStepBehavior.create(this.designerContext, newStep));
 			return true;
 		}
 		return false;
 	}
-}
 
-function createStep(step: StepDefinition): Step {
-	const newStep = ObjectCloner.deepClone(step) as Step;
-	newStep.id = Uid.next();
-	return newStep;
+	private activateStep(step: StepDefinition): Step {
+		const newStep = ObjectCloner.deepClone(step) as Step;
+		newStep.id = this.uidGenerator ? this.uidGenerator() : Uid.next();
+		return newStep;
+	}
+
+	private getConfiguration(): ToolboxConfiguration {
+		if (!this.configuration) {
+			throw new Error('Toolbox is disabled');
+		}
+		return this.configuration;
+	}
 }